requirements.txt
streamlit>=1.36
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.26